+++ rest/src/main/java/com/wherecar/rest/car/presentation/CarController.java
@@ -5,7 +5,7 @@ import com.wherecar.rest.car.application.dto.CarOverviewResponse;
 import com.wherecar.rest.car.application.dto.CarResponse;
 import com.wherecar.rest.car.application.dto.CarRegisterRequest;
 import com.wherecar.rest.car.application.CarService;
-import com.wherecar.rest.common.auth.AuthUtil;
+import com.wherecar.rest.security.auth.AuthUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
